Added Skills and min.js
I added the tables of skills and the skills section. I also added a min
version of the idea generator js. I formatted the CSS and JS files to
use proper spacing. There is a picture placeholder in the "about"
section. I also cleaned up the JS code to use easier to read variables,
and added plural languages.